+++ Changelog.md
@@ -7,12 +7,16 @@ Breaking changes
 New
 
 * Added an optional push size limit to `MessageBuilder`. ([#348])
+* Added `FromStr` impls for `Rcode` and `OptRcode`. ([#357])
+* Added `OptRcode::is_ext` to check if the code is an extended code.
+  ([#358])
 
 Bug fixes
 
 * Fixed a mistake in the tsig module while calculating the start of
   the TSIG record when there were other records in the additional section,
   causing the TSIG code to fail if OPT records were in use. ([#333])
+* Fixed the mnemonic for the `NOTAUTH` rcode – it was `NOAUTH`. ([#360])
 
 Unstable features
 
@@ -34,6 +38,9 @@ Other changes
 [#336]: https://github.com/NLnetLabs/domain/pull/336
 [#338]: https://github.com/NLnetLabs/domain/pull/338
 [#348]: https://github.com/NLnetLabs/domain/pull/348
+[#357]: https://github.com/NLnetLabs/domain/pull/357
+[#358]: https://github.com/NLnetLabs/domain/pull/358
+[#360]: https://github.com/NLnetLabs/domain/pull/360
 
 ## 0.10.1